Removes exiting on network issue, adds up metric
We've had incidents where the connection to pingdom flakes out,
and the exporter exits. It then retries a few times, but
ultimately fleet kills it, and this wakes an operator via
prometheus alerts. This changeset removes the exiting on the
network issue, and instead modifies an up metric, which we can
add finer grained alerting on.