Update to use Golang 0.12, and modified the docker image build.
Binary is now built with the latest golang, and the stale dependency
management entries (from old golang tools) were removed.

Updated the docker image build process to build the binary inside
docker, so building an image alone does not leave any artifacts other
than the image.

Updated the build to apply the build VERSION as a tag on the docker
image.  This required changing the VERSION format slightly, removing the
plus sign.  Instead of 0.1.1+git the VERSION format is now
0.1.1-<git-sha>.  If the '-git' is removed from the VERSION file, the
image tag will just be the simple version.